Create edit authentication and redirect
Previously, anyone with the right url and the user id could access the edit page of an account. A feature has been implemented to have only the right user to be able to access the edit page of that particular page. This commmit should have been broken into smaller pieces. Another redirect is done for friendly forwarding. More details can be found on Chapter 10 of railstutorial.com